+++ Changelog.md
@@ -10,6 +10,7 @@
 
 ## FF7
 
+- Core: Fix `ERROR: COULD NOT CHANGE TO DIRECTORY` error when starting the Condor minigame
 - Input: Allow Cloud to walk/run based on the left analogue stick position ( https://github.com/julianxhokaxhiu/FFNx/issues/523 )
 
 ## FF8